Added multiprocessing to speed up the simulation.
I guess it worked, but was it worth my sanity?
Multiprocessing doesn't work in Jupyter Notebooks unless the target is imported from a regular python script
"Multiprocess" is supposed to fix that but doesn't capture globals
Setting dill to recursive is supposed to fix that but it screws me on Windows
I moved the simulation code to a regular python script and had to hack together some bullshit so that it doesn't need to import any notebooks
I may cry.